Restructure tests by database
