+++ tests/test_helpers.py
@@ -46,56 +46,12 @@ def assert_dashboard(browser: mechanicalsoup.StatefulBrowser):
     assert "dashboard" in page.text.lower()
     assert "about this system" in page.text.lower()
 
-def install_sqlite(browser: mechanicalsoup.StatefulBrowser):
-    setup_admin_password(browser)
-
+def assert_upgrade(browser: mechanicalsoup.StatefulBrowser):
+    browser.open(BASE_URL)
     page = browser.get_current_page()
-    assert "baïkal database setup" in page.text.lower()
-    
-    browser.select_form("form")
-    browser.submit_selected()
-
-    assert_installed(browser)
-    assert_dashboard(browser)
-    
-def install_pgsql(browser: mechanicalsoup.StatefulBrowser):
-    setup_admin_password(browser)
-
+    assert "baïkal upgrade wizard" in page.text.lower()
+    browser.follow_link(text="Start upgrade")
     page = browser.get_current_page()
-    assert "baïkal database setup" in page.text.lower()
-    browser.select_form("form")
-    browser["data[backend]"] = "pgsql"
-    browser.submit_selected()
-
-    page = browser.get_current_page()
-    assert "postgresql host" in page.text.lower()
-    browser.select_form("form")
-    browser["data[pgsql_host]"] = "127.0.0.1"
-    browser["data[pgsql_dbname]"] = "baikal_test"
-    browser["data[pgsql_username]"] = "baikal"
-    browser["data[pgsql_password]"] = "baikal"
-    browser.submit_selected()
-
-    assert_installed(browser)
-    assert_dashboard(browser)
-
-def install_mysql(browser: mechanicalsoup.StatefulBrowser):
-    setup_admin_password(browser)
-
-    page = browser.get_current_page()
-    assert "baïkal database setup" in page.text.lower()
-    browser.select_form("form")
-    browser["data[backend]"] = "mysql"
-    browser.submit_selected()
-
-    page = browser.get_current_page()
-    assert "mysql host" in page.text.lower()
-    browser.select_form("form")
-    browser["data[mysql_host]"] = "127.0.0.1"
-    browser["data[mysql_dbname]"] = "baikal_test"
-    browser["data[mysql_username]"] = "baikal"
-    browser["data[mysql_password]"] = "baikal"
-    browser.submit_selected()
-
-    assert_installed(browser)
+    assert "baïkal has been successfully upgraded" in page.text.lower()
+    browser.follow_link(text="Access the Baïkal admin")
     assert_dashboard(browser)
